manual_contol_ssetpoint: fix mode slot numbering
This fix is necssary because usually uORB structs get initialized with
all zeros and then et filled with the fields that are actually in use.
The number 0 for the mode slot was already commanding to switch to the
mode in slot one even though for example the joystick input via mavlink
does not use this mechanism at all.